migrate colour boxes into nclx specific file
