+++ libheif/pixelimage.h
@@ -24,7 +24,7 @@
 
 #include "heif.h"
 #include "error.h"
-#include "box.h" // only for color_profile, TODO: maybe move the color_profiles to its own header
+#include "nclx.h"
 
 #include <vector>
 #include <memory>
